Abort if the command for `run` fails and `exit_on_failure?` is true
According to the description of `exit_on_failure?`,
the process should exit with status 1 if any error happens.
I think that it should be also applied to `run`.

Although we can pass `abort_on_failure` option for `run`,
passing it every time seems a little redundant when we call
`run` multiple times. This is the other reason for this change.